feat: pass the CB func via props to update the parent state and force an re render of all child components
Passes the CB that will trigger an update of the book shelfs, both in the FE and BE - through the API call -, and that will update the values of all child components. By updating the BE by calling the update() endpoint, we make sure than when the page refreshes the information persists.